requirements.txt
im-squonk2-client >= 1.13.3, < 2.0.0
python-dateutil == 2.8.2
rich == 12.5.1
pyyaml == 6.0
